refactor: Split LoanInvariant into LoanBrokerInvariant and LoanInvariant
Separate ValidLoanBroker and ValidLoan into their own header/source
file pairs, following the existing pattern of other invariant files
(e.g. AMMInvariant, VaultInvariant).
